# docs/swagger.yaml
basePath: /
definitions:
  api.buildRequest:
    properties:
      commit_hash:
        description: CommitHash pins the build to a specific commit (also used to
          name the Job).
        type: string
      namespace:
        description: Namespace to run the build Job in (defaults to "default").
        type: string
      repo_url:
        description: RepoURL is the git repository to clone and build.
        type: string
    type: object
  api.createAppRequest:
    properties:
      app_name:
        description: e.g. "my-laravel-app"
        type: string
      commit_hash:
        description: e.g. "latest"
        type: string
      namespace:
        description: defaults to "default"
        type: string
      replicas:
        description: defaults to 1
        type: integer
      repo_url:
        description: e.g. "https://github.com/laravel/laravel.git"
        type: string
    type: object
  api.response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a simple PaaS API for Laravel applications on Kubernetes.
  title: my-paas API
  version: "1.0"
paths:
  /apps:
    get:
      description: List all Laravel deployments in a namespace.
      parameters:
      - description: Namespace (defaults to 'default')
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.response'
      summary: List all apps
      tags:
      - apps
    post:
      consumes:
      - application/json
      description: Create a new Laravel deployment and trigger initial build.
      parameters:
      - description: App details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createAppRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.response'
      summary: Create a new Laravel app
      tags:
      - apps
  /apps/{name}:
    delete:
      description: Delete an app's deployment.
      parameters:
      - description: App name
        in: path
        name: name
        required: true
        type: string
      - description: Namespace (defaults to 'default')
        in: query
        name: namespace
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.response'
      summary: Delete an app
      tags:
      - apps
    get:
      description: Get the status of a single app.
      parameters:
      - description: App name
        in: path
        name: name
        required: true
        type: string
      - description: Namespace (defaults to 'default')
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.response'
      summary: Get app status
      tags:
      - apps
  /apps/{name}/build:
    post:
      consumes:
      - application/json
      description: Clone, build Docker image, push, then roll the deployment.
      parameters:
      - description: App name
        in: path
        name: name
        required: true
        type: string
      - description: Build details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.buildRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.response'
      summary: Trigger a new build
      tags:
      - builds
swagger: "2.0"